+++ src/ast/test.rs
@@ -92,7 +92,7 @@ fn test_diag_expect(comp: (&String, &Vec<super::diag::PLDiag>)) {
     expected.assert_eq(&format!("{:#?}", diag));
 }
 
-fn sanitize_diag(diag: &Vec<super::diag::PLDiag>) -> Vec<super::diag::PLDiag> {
+fn sanitize_diag(diag: &[super::diag::PLDiag]) -> Vec<super::diag::PLDiag> {
     let mut diag = diag
         .iter()
         .map(|d| {
